Fix timeout, and move cli/add/binding.go to cli/add/binding/binding.go
