+++ pkg/odo/cli/add/binding/binding.go
@@ -1,4 +1,4 @@
-package add
+package binding
 
 import (
 	"context"
